bindep.txt: Don't install ruby-devel on centos 7/8
vagrant is not available in Centos 7 and 8, so the tools/test-setup.sh
script is installing the rpm from Vagrant upstream.
This rpm comes with its own bundled version of ruby and it's different
than the one in Centos 7/8. In the end, it will compile for the Centos
version of ruby, but vagrant will use the embedded version. This results
in:
in `require': incompatible library version - /home/zuul/.vagrant.d/gems/2.4.9/gems/ruby-libvirt-0.7.1/lib/_libvirt.so (LoadError)

The easiest way to fix this is to not install ruby-devel on centos 7/8.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>