test-setup.sh: Fix building on Centos 7
After fixing the ruby-devel interaction with the bundled ruby
with vagrant, the build still fails on Centos 7, due to link
failures during ruby-libvirt build.
To sum up, the problem is between bundled openssl with vagrant and
system openssl (missing symbols).

In the vagrant-libvirt plugin installation page, there's a mention
of the command line to use in this case. This commit change the plugin
compilation to detect the vagrant rpm, by looking for /opt/vagrant/bin/vagrant
and use the needed variables to get the build working.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>