molecule_vagrant/modules/vagrant.py: Fix vagrant working directory
vagrant does a lot of thing according to its working directory, like
reading the Vagrantfile, launching external commands, etc...
Since vagrant is started by python-vagrant, itself started by ansible
in the directory containing the create/destroy/... test playbooks, it means
that the current directory is something like $PYTHONDIR/molecule_vagrant/playbooks/
and not the molecule ephemeral directory.

In order to make sure that vagrant finds its files, the code is using
VAGRANT_CWD environment variable, which allows vagrant to work but it
doesn't change the current directory. For libvirt, it's not a problem but
for virtualbox, it will create files in this current directory. Writing
to this directory may even be forbidden.

Change the code to set the 'root' parameter of the vagrant class,
which will chdir() to the molecule ephemeral directory.

For the test case, the idea was to find a way to get vagrant current
directory. The Vagrantfile is a ruby file parsed and interpreted by
vagrant, so at that time, the current directory is the one of vagrant.
This value computed with 'Dir.pwd' in the Vagrantfile, is stored in a
file inside the instance with the shell provisioner and the ansible
verifier checks that its content is the molecule ephemeral directory.

Fixes: #20
Signed-off-by: Arnaud Patard <apatard@hupstream.com>