molecule_vagrant/modules/vagrant.py: Modify the Vagrant{file,.yml} only at creation
The current code is rewriting at every stage the Vagrantfile and
vagrant.yml files. At VM creation stage, it's fine but when destroying,
since the platform_box parameter is optional, this will result in invalid
vagrant.yml: the config.vm.box option will be empty.
Instead of making the platform_box mandatory in creation and halt/destroy
cases, make it mandatory for the VM creation and create/modify only
at this step.

This will possibly fix bug 19 and superseed PR 25 but this has to be
confirmed by the submitter.

The side effects of this commit are:
- it will break the setup of multi node. Multi node support gives the feeling that it's working,
but it reality, it is not. What's happening is that the configuration file is overwritten when
creating/destroying the different nodes (instead of having one Vagrantfile for every
node).
- it will probably conflict with PR 31.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>