bindep.txt: Fix support for dpkg platforms
- remove "build-dep" dependency as it doesn't exist.
  Replace it with build-essential, even if the idea was probably to
  use "apt-get build-dep vagrant ruby-libvirt" as it the vagrant-libvirt doc
- Install python3/python3-dev
- Install qemu-system/qemu-utils.

This has been tested on debian 10 (with Debian's official vagrant
and vagrant-libvirt package) and on Xenial with hashicorp debian
package and manually installed vagrant-libvirt.

The only remaining possible test to do would be to try installing
vagrant-libvirt manually and using the offician Debian/Ubuntu vagrant
packages but I'm not sure it's a case needed for now.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>