tools/test-setup.sh: Fix vagrant command line
The usermod command line is run on the same session as the one running
vagrant. This means that the current user is not in the libvirt group
when vagrant is called. It should be failing, as in Xenial, but the
libvirt socket rights are 0777 on Debian buster / Centos 8 / Fedora 31
which means that the bug was not noticed before.

Since I don't want to use exec -l (in order to get the "vagrant destroy -f"
command), I've choosen to use "sudo su -l "$(whoami)" -c " instead. Not
really nice but works.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>