molecule_vagrant/test/: Allow to set vagrant box to use for tests
Currently, the vagrant boxes used can't be changed and are not all
the same, which means that during testing, different boxes are downloaded
and used.
As a solution, allow to set a box for every box, thanks to the new
TESTBOX environment variable. If not set, the current behaviour is not
changed.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>