tools/test-setup.sh: Create a vagrant box from setup test image
The default image used by molecule-vagrant and the test-setup.sh
script are using the generic/alpine310 box. This box can be used
for testing molecule-vagrant too but using it will mean to download
python3 and rsync everytime a new test is started.
To avoid doing it, create a new vagrant box out of the freshly
started VM and using Vagrantfile provisionning to install python3
and rsync.
To avoid having to handle the vagrant insecure ssh key pubkey, set
config.ssh.insert_key to false.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>