molecule_vagrant/test/functional/test_func.py: handle tcg+alpine on bionic
The libvirt is defaulting to EPYC processor inside the virtual machine,
even if it's more or less broken on the qemu tcg found in bionic. This
leads to the boot stuck at "ssh-keygen -A". As a workaround, use centos/7
inside the VM for this specific case. It'll be slow but it's better than
not working testsuite.

I could have switched to core2duo, as done for the tools/Vagrantfile file,
but changing the box was easier.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>