tools/test-setup.sh: Enable debug to vagrant
The CI is still failing to get an IP from time to time. According to the bug I found,
switching to e1000 should have been enough, but looks like it's not.
Given that I can reproduce it locally, add some debug in hope this will allow
me to find the issue.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>