molecule_vagrant/modules/vagrant.py: Stop depending on molecule.config
Since molecule commit "Avoid parallel execution of a specific schenario"
(65ccd6b), the molecule scenario directory is protected by a lock.
Which is fine. Except that the vagrant module is using molecule.config.Config
to get:
- the molecule ephemeral directory,
- path to the Vagrantfile.

This will result in the vagrant module trying again to get the lock help
by molecule and fails with messages like:
Retrying to acquire lock on /home/zuul/.cache/molecule/scenarios/config_options, waiting for 30 seconds
Retrying to acquire lock on /home/zuul/.cache/molecule/scenarios/config_options, waiting for 60 seconds
Retrying to acquire lock on /home/zuul/.cache/molecule/scenarios/config_options, waiting for 90 seconds
Retrying to acquire lock on /home/zuul/.cache/molecule/scenarios/config_options, waiting for 120 seconds
Timedout trying to acquire lock on /home/zuul/.cache/molecule/scenarios/config_options

As a solution, add a new parameter to the module, called 'workdir', used
to set the vagrant working directory, which is in reality what's the
path used by the module to store its files and this path is obviously
the molecule ephemeral directory in case of molecule.
So, if this parameter is not set, it will defaults to the molecule
ephemeral directory.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>