molecule_vagrant/modules/vagrant.py: Fix fallback
The functions up/destroy/halt of the VagrantClient class
are exiting with proper fail/exit_json but in some cases,
the last line of main() may still be reached but leads to an
other error since there's no module.result.
While it's not sure how this code is reached, at least fix
that broken line by failing and returning VagrantClient class
result member.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>