tools/test-setup.sh: Fix libvirt access
Most distributions relies on unix rights to configure
libvirt socket access, so the user has to be added to the 'libvirt'
group to use it. The challenge is that it implies login/logout.

The workaround to use newgrp is failing since it's spawning a new
shell, so the group change is not applied to current shell.

As a solution, use policykit to handle the access rights to the
unix socket. It's the same thing as for unix rights, except that
policykit doesn't rely on the current shell, so no need to login/logout
to get the group change applied.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>