molecule_vagrant/modules/vagrant.py: Set cpu to qemu64 with qemu driver
When TCG is used instead of KVM, libvirt defaults to using the EPYC
cpu model on some systems (even if the host is an intel cpu). This
leads to some weird issues, like ssh-keygen -A hanging at the
boot of alpine VM in the CI.
In previous works, I've work arounded the issue by switch the VM to
using centos instead of alpine but I think it's a good idea to
provide good working default, so the vCPU model is set to qemu64
when TCG is used, if the vCPU model is not specified.
The tools/Vagrantfile has been updated to match this vCPU model too.

Side-note: I'm still wondering if the same thing should be done for
systems without KVM, aka automatically default to TCG in this case.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>