Improve test box handling (#62)
* molecule_vagrant/test/: Allow to set vagrant box to use for tests

Currently, the vagrant boxes used can't be changed and are not all
the same, which means that during testing, different boxes are downloaded
and used.
As a solution, allow to set a box for every box, thanks to the new
TESTBOX environment variable. If not set, the current behaviour is not
changed.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>

* tools/test-setup.sh: Create a vagrant box from setup test image

The default image used by molecule-vagrant and the test-setup.sh
script are using the generic/alpine310 box. This box can be used
for testing molecule-vagrant too but using it will mean to download
python3 and rsync everytime a new test is started.
To avoid doing it, create a new vagrant box out of the freshly
started VM and using Vagrantfile provisionning to install python3
and rsync.
To avoid having to handle the vagrant insecure ssh key pubkey, set
config.ssh.insert_key to false.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>

* tox.ini: Use newly created test box

Use the vagrant box created with test-setup.sh to do the testing.
This should reduce testing time.

Since github actions are not using tools/test-setup.sh, configure
the tests to use generic/alpine310 in this case.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>