.github/workflows/tox.yml: Create testbox too
For unknown reasons, things have worked so far in github action
with the alpine box, but got recently a failure due to missing
python (only on 1 of the several scenarii in the CI run). So,
make sure it's there by using the testbox box.

In order to do that, I've moved the box creation out of test-setup.sh
into a new script, used by test-setup.sh and github action.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>