molecule_vagrant/modules/vagrant.py: don't use boolean jinja test
On one of my test env, got a failure from jinja:
The error was: jinja2.exceptions.TemplateAssertionError: no test named 'boolean'

So, check that the value is same as true or false instead of testing
'boolean' type.

I'm not exactly sure where the culprit is but the change should be harmless.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>