molecule_vagrant/playbooks/prepare.yml: Better error handling
Currently, if python installation fails, no error or message is
returned and ansible will fail to execute following tasks since
python is not there.
As a solution, make the raw: command file and retry it 3 times
in case of failure. To keep current behaviour, ignore errors.

With that, it should be easier to spot installation errors in
the logs and python bootstrap should be more robust.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>