tools/Vagrantfile: Reduce VM memory to 512
Default alpine box setting is to require 2048M, which seems to be too much
for Zuul 1G.
After local testing, the molecule-vagrant default setting of 512M is enough
to work inside a VM with 1G of RAM. So, reduce the memory usage to 512M for
the VM used to create the test box. If this fails due to memory pressure,
testing will fail later too.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>