zuul.d/layout.yaml: Set molecule-vagrant-fedora parent to tox-py38
Since https://github.com/ansible/ansible-zuul-jobs/commit/472249173a10a53e6924f4df772cd993bce439e4,
the ansible-tox-py38 job is relying on centos-8-stream and tries to
install python38-devel.
Here, we're using fedora in order to test molecule-vagrant with
distribution packages and libvirt. And trying to install python38-devel
is failing since there's no python38-devel.

While I'm going to create a branch in my fork of with fixes for centos-8-stream,
it's a pain to use since it needs the workaround for krb5 and a new
workaround for libssh. The real fix would be hashicorp to fix their broken
RPMs but looks like months (years?) and bug reports are not enough
to get them fix this.
The other possibility would have been to use fedora and py39 but there's
no ansible-tox-py39 ATM.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>