fix: use the instance name as the iterator
in the vm.provider block. When molecule provider.name == 'vagrant-vmware-desktop'
vagrant takes exception to the "-" in the iterator name, and the syntax check fails.
Since this is a per instance configuration, using the instance name as the iterator
seems sensible. Although can still have - in it, at least its something that
the molecule user can change.