molecule_vagrant/modules/vagrant.py: Use vagrant.validate()
python-vagrant 1.0.0 adds a new validate() function, so use it
instead of the open coded version here. [ So, bump minimal version
to 1.0.0 ].

While at it, add a small test case for it to make sure it's handled
properly as I managed to get it wrong in my first version of this
patch.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>