Allow to force vagrant provider in vagrant up
Currently, vagrant is still choosing the provider to use on its own
liking. While it's fine for cases like our CI or people who are fine
with default VM configurations, it's not always the case.

See:
- https://github.com/ansible-community/molecule-vagrant/pull/174
- https://github.com/ansible-community/molecule-vagrant/issues/181

So, add a new configuration setting to call vagrant up with
--provider, in order to force the provider to use and lead to an
error if it's not usable. The default value is to keep current
behaviour to not break things.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>