molecule_vagrant/modules/vagrant.py: Allow to define per-providers options
In most test cases, there's only one provider to handle (being vbox, libvirt, ...)
but in some cases, one may want to run the CI against several providers
like in our CI with libvirt (my local testing, zuul) and gha (vbox). In these
cases, either create different scenario per provider or try to use
"clever" things like what's done in our CI with the network scenario.

A better solution is to allow to define per-providers options, as done in a
lot of Vagrantfiles like this:

Vagrant.configure("2") do |config|
  ...
  config.vm.provider :libvirt do |prov|
    ...
  end
  config.vm.provider :virtualbox do |prov|
    ...
  end
...
end

This will make things cleaner and avoid such "hacks" as in the one in the CI
scenario.

This new setting (providers_options) is meant to extend current provider_option
setting and not to replace it.

This means:
- current behaviour is unchanged,
- the provider_* settings are applied to all providers configured with
  providers_options.

Moreover, the memory and cpus settings will be applied to all providers specified
in providers_options.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>