Pass --provider to vagrant when provider name specified
Currently, vagrant is still choosing the provider to use on its own
liking. This may cause troubles like VM memory/cpus or provider
options ignored if the provider use is not the one specified.

See:

- https://github.com/ansible-community/molecule-vagrant/pull/174
- https://github.com/ansible-community/molecule-vagrant/issues/181

So, pass --provider when the provider is set in the molecule.yml.
If the test doesn't care about the provider, it should remove the
provider name setting in the molecule.yml.

Note: While this may break some setup, I believe this will prevent
new reports/issues like the one previously noted. In case one
molecule.yml file for multiple providers is wanted, use the new
providers_options, as done in our CI scenario.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>